+++ R/dev-help.R
@@ -36,7 +36,13 @@ dev_help <- function(topic,
 
   loc <- dev_topic_find(topic, dev_packages)
 
-  if (is.null(loc$path)) {
+  if (!is.null(loc$path) && !fs::file_exists(loc$path)) {
+    # Documentation topic might have moved, so reset topic index and try again
+    dev_topic_index_reset(loc$pkg)
+    loc <- dev_topic_find(topic, dev_packages)
+  }
+
+  if (is.null(loc$path) || !fs::file_exists(loc$path)) {
     cli::cli_abort("Can't find development topic {.arg {topic}}.")
   }