+++ tests/testthat/test-help.R
@@ -72,6 +72,28 @@ test_that("show_help and shim_question files for devtools-loaded packages", {
     ))
 })
 
+test_that("shim_help and shim_questions works if topic moves", {
+  load_all(test_path('testHelp'))
+  on.exit(unload(test_path('testHelp')))
+
+  path_man <- test_path("testHelp/man/")
+  rel_rd_path <- function(x) {
+    as.character(fs::path_rel(x$path, path_man))
+  }
+
+  expect_equal(rel_rd_path(shim_help("foofoo")), "foofoo.Rd")
+  expect_equal(rel_rd_path(shim_question("foofoo")), "foofoo.Rd")
+
+  fs::file_move(fs::path(path_man, "foofoo.Rd"), fs::path(path_man, "barbar.Rd"))
+  on.exit(
+    fs::file_move(fs::path(path_man, "barbar.Rd"), fs::path(path_man, "foofoo.Rd")),
+    add = TRUE
+  )
+
+  expect_equal(rel_rd_path(shim_help("foofoo")), "barbar.Rd")
+  expect_equal(rel_rd_path(shim_question("foofoo")), "barbar.Rd")
+})
+
 test_that("dev_help works with package and function help with the same name", {
   load_all(test_path('testHelp'))
   on.exit(unload(test_path('testHelp')))